requirements.txt
aioredis==1.3.1
amqp==2.6.1
appdirs==1.4.3
arabic-reshaper==2.1.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.4.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
charset-normalizer==2.0.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-cors-headers==3.7.0
django-dbtunnel==1.0.0
django-filter==2.4.0
django-redis==5.0.0
django-rest-swagger==2.2.0
django-webpack-loader==0.7.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.7.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
et-xmlfile==1.1.0
future==0.18.2
hiredis==2.0.0
html5lib==1.0.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
kafka-python==2.0.2
kombu==4.6.11
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.21.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==20.3
pandas==1.2.5
paramiko==2.7.2
pep517==0.8.2
Pillow==8.3.2
progress==1.5
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.5
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.6.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
service-identity==18.1.0
simplejson==3.17.3
six==1.14.0
sqlparse==0.4.1
sshtunnel==0.4.0
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
zope.interface==5.4.0
